MNI bug fixed
FIxed the bug that returned incorrect areas for the MNI mapping of the electrodes.

Error was due to some contacts being in areas without label, and automatically assigned to be "Optic Chiasm". Now, if a contact is in an empty label, surrounding voxels (up to 5 iterations) are checked until a label (or multiple) is found. The majority of surrounding labels are selected as possible mapping of that contact. If no labels are found, the contact is assigned to an "unknown" region.