bug fixed
There was a bug that crashed the code. Not located yet, but went to a previous version of the code to solve it. All functions in the previous commit are kept, the only difference (not considering the bug) is that there are commented lines in the script of old code not used anymore.