Warning pop-up + aesthetic issues
If the user makes a mistake and the markup list does not contain the start and end of each electrode (f.e. the user writes P1 for the tip but PI12 for the end) a warning message arises (as a pop-up window) letting the user know about it.

Changes in the location of the "Transfer" option in the GUI so it is not clicked by mistake. Unabled by default.