two-point udate
Although that for contact generation the placement  of just two contacts were required, there was a dependency of the end of the shaft to be included by the user. Now the end of the shaft is generated automatically and the input markup list should just contain the first and last contacts of each electrode. And bending points in case that is necessary.